+++ src/algo.java
@@ -1,8 +1,8 @@
 
 public class algo {
 
-	public static void f(long a) {
-		long n=a;
+	public static void f(double a) {
+		double n=a;
 		while(n>0) {
 			n--;
 		}
